+++ src/components/Sections/Top/SecuredUsersCounter.svelte
